% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_depth_bin.R
\name{agg_depth_bin}
\alias{agg_depth_bin}
\title{Aggregate depth bins}
\usage{
agg_depth_bin(this.glorys, agg.lev, max.level)
}
\arguments{
\item{this.glorys}{}

\item{agg.lev}{}

\item{max.level}{}
}
\description{
Aggregate depth bins
}
