% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_clim.R
\name{cal_clim}
\alias{cal_clim}
\title{Calculate climatology and delta anomalies}
\usage{
cal_clim(this.rel, depth.hist.files, depth.future.files)
}
\arguments{
\item{this.rel}{}

\item{merged.hist.files}{}

\item{merged.future.files}{}
}
\description{
Calculate climatology and delta anomalies
}
