% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-catchdata.R
\docType{data}
\name{catchdata}
\alias{catchdata}
\title{DATASET_TITLE}
\format{
A data frame with 1338506 rows and 8 variables:
\describe{
\item{\code{rnp_code}}{character COLUMN_DESCRIPTION}
\item{\code{NOM_ENT}}{character COLUMN_DESCRIPTION}
\item{\code{NOM_LOC}}{character COLUMN_DESCRIPTION}
\item{\code{fishery_office}}{character COLUMN_DESCRIPTION}
\item{\code{species}}{character COLUMN_DESCRIPTION}
\item{\code{landed_w_kg}}{double COLUMN_DESCRIPTION}
\item{\code{value_mxn}}{double COLUMN_DESCRIPTION}
\item{\code{year}}{double COLUMN_DESCRIPTION}
}
}
\usage{
catchdata
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
