% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_catch.R
\name{clean_catch}
\alias{clean_catch}
\title{Combine catch files}
\usage{
clean_catch(this.catchfileno, catch.files.clean)
}
\arguments{
\item{this.catchfileno}{}
}
\value{
catch.file
}
\description{
Combine catch files
}
