% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_locs_pesca.R
\name{clean_locs_pesca}
\alias{clean_locs_pesca}
\title{Import and prepare list of fishery names}
\usage{
clean_locs_pesca()
}
\value{
locs.pesca
}
\description{
Import and prepare list of fishery names
}
