% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_grid.R
\name{correct_grid}
\alias{correct_grid}
\title{regrid files to assign lat lon dimensions}
\usage{
correct_grid(this.file, this.var)
}
\arguments{
\item{this.file}{}

\item{this.var}{}

\item{weight.dir}{}
}
\description{
regrid files to assign lat lon dimensions
}
