% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georef_catch.R
\name{georef_catch}
\alias{georef_catch}
\title{Georeferrence list of catch locations}
\usage{
georef_catch(catchlocations, georeflocskm)
}
\arguments{
\item{catchlocations}{}

\item{georeflocskm}{}
}
\value{
georefcatchlocs
}
\description{
Georeferrence list of catch locations
}
