% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georef_catch_data.R
\name{georef_catch_data}
\alias{georef_catch_data}
\title{Georeferrence all catch data}
\usage{
georef_catch_data(georefcatchlocs, catchdata)
}
\arguments{
\item{georefcatchlocs}{}

\item{catchdata}{}
}
\value{
catch.plot Plot of catch locations in the Gulf of California and point in the Pacific that were eliminated
}
\description{
Georeferrence all catch data
}
