% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georeff_catch.R
\name{georref_catch}
\alias{georref_catch}
\title{Georeferrence list of catch locations}
\usage{
georref_catch(catchlocations, georeflocskm)
}
\arguments{
\item{catchlocations}{}

\item{georeflocskm}{}
}
\value{
georeffcatchlocs
}
\description{
Georeferrence list of catch locations
}
