% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catch.R
\name{get_catch}
\alias{get_catch}
\title{Get and stadarize catch records}
\usage{
get_catch(this.catchfileno, catch.files, coast.selecc, locs.pesca)
}
\arguments{
\item{this.catchfileno}{}

\item{catch.files}{}

\item{coast.selecc}{}

\item{locs.pesca}{}
}
\value{
this.permit.corr
}
\description{
Get and stadarize catch records
}
