% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climatologyglorys.R
\name{get_climatologyglorys}
\alias{get_climatologyglorys}
\title{Get GLORYS climatology data}
\usage{
get_climatologyglorys(min_lon, max_lon, min_lat, max_lat, out_dir, data.set.id)
}
\arguments{
\item{min_lon}{}

\item{max_lon}{}

\item{min_lat}{}

\item{max_lat}{}

\item{out_dir}{}

\item{data.set.id}{}
}
\description{
Get GLORYS climatology data
}
