% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distance.R
\name{get_distance}
\alias{get_distance}
\title{Get distance to nearest fishery office}
\usage{
get_distance(thisloc, fishery.offices)
}
\arguments{
\item{thisloc}{}

\item{fishery.offices}{}
}
\description{
Get distance to nearest fishery office
}
