% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_futuretemp.R
\name{get_futuretemp}
\alias{get_futuretemp}
\title{Get future mean temperature for cluster areas}
\usage{
get_futuretemp(
  eachrel,
  rel.list,
  future.files,
  future.time,
  future.path,
  cluster.nums,
  cluster.coords,
  ini.depth,
  end.depth
)
}
\arguments{
\item{rel.list}{}

\item{future.files}{}

\item{future.time}{}

\item{future.path}{}

\item{cluster.nums}{}

\item{cluster.coords}{}
}
\description{
Get future mean temperature for cluster areas
}
