% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_glorysdata.R
\name{get_glorysdata}
\alias{get_glorysdata}
\title{Get monthly GLORYS data for each file}
\usage{
get_glorysdata(
  eachglorysfile,
  glorys.files,
  this.ocean.grid,
  this.catch.geo,
  this.depth,
  list.tempvalues
)
}
\arguments{
\item{eachglorysfile}{}

\item{glorys.files}{}

\item{this.ocean.grid}{}

\item{this.catch.geo}{}

\item{this.depth}{}

\item{list.tempvalues}{}
}
\value{
list.tempvalues
}
\description{
Get monthly GLORYS data for each file
}
