% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locations_catch.R
\name{get_locations_catch}
\alias{get_locations_catch}
\title{Get fishery office from catch records}
\usage{
get_locations_catch(catchdata)
}
\arguments{
\item{catchdata}{timeseries of catch for Gulf of California states}
}
\value{
catch.locations
}
\description{
Get fishery office from catch records
}
