% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meanclustertemp.R
\name{get_meanclustertemp}
\alias{get_meanclustertemp}
\title{Get temperature by cluster}
\usage{
get_meanclustertemp(
  this.time.interval,
  glorys.files,
  cluster.locs,
  num.clusters,
  time.interval,
  start.time.interval,
  out.dir
)
}
\arguments{
\item{this.time.interval}{}

\item{glorys.files}{}

\item{num.clusters}{}

\item{time.interval}{}

\item{start.time.interval}{}

\item{out.dir}{}
}
\description{
Get temperature by cluster
}
