% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meandepth.R
\name{get_meandepth}
\alias{get_meandepth}
\title{Get mean depth data}
\usage{
get_meandepth(this.file, depthname)
}
\arguments{
\item{this.file}{}

\item{depthname}{description}
}
\value{
this.file.data
}
\description{
Get mean depth data
}
