% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_monthlyglorys.R
\name{get_monthlyglorys}
\alias{get_monthlyglorys}
\title{Get monthly GLORYS data}
\usage{
get_monthlyglorys(
  thisdateeentry,
  start.time.interval,
  end.time.interval,
  min_lon,
  max_lon,
  min_lat,
  max_lat,
  out_dir,
  data.set.id
)
}
\arguments{
\item{thisdateeentry}{}

\item{start.time.interval}{}

\item{end.time.interval}{}

\item{min_lon}{}

\item{max_lon}{}

\item{min_lat}{}

\item{max_lat}{}

\item{out_dir}{}

\item{data.set.id}{}
}
\description{
Get monthly GLORYS data
}
