% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nc_description.R
\name{get_nc_description}
\alias{get_nc_description}
\title{Get raster resolution, description and plot}
\usage{
get_nc_description(thisraster)
}
\arguments{
\item{thisraster}{}
}
\value{
this.terra.rast terra raster
}
\description{
Get raster resolution, description and plot
}
