% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pointdistance.R
\name{get_pointdistance}
\alias{get_pointdistance}
\title{Title Get distance from catch to neares grid polygons}
\usage{
get_pointdistance(eachcatchpoint, catch.pacific.spatial, gc.ocean.grid)
}
\arguments{
\item{eachcatchpoint}{}

\item{catch.pacific.spatial}{}

\item{gc.ocean.grid}{}
}
\value{
this.ocean.grid
}
\description{
Title Get distance from catch to neares grid polygons
}
