% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pointtemp.R
\name{get_pointtemp}
\alias{get_pointtemp}
\title{Get temperature for catch points}
\usage{
get_pointtemp(
  eachcatchpoint,
  this.depth,
  catch.pacific.spatial,
  gc.ocean.grid,
  glorys.files
)
}
\arguments{
\item{eachcatchpoint}{}

\item{this.depth}{}

\item{catch.pacific.spatial}{}

\item{gc.ocean.grid}{}

\item{glorys.files}{}
}
\description{
Get temperature for catch points
}
