% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperslab_depth.R
\name{hyperslab_depth}
\alias{hyperslab_depth}
\title{Hyperslab depth layers}
\usage{
hyperslab_depth(this.merge.file, ini.depth, end.depth)
}
\arguments{
\item{this.merge.file}{}

\item{ini.depth}{}

\item{end.depth}{}
}
\description{
Hyperslab depth layers
}
