% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperslab_spatial_area.R
\name{hyperslab_spatial_area}
\alias{hyperslab_spatial_area}
\title{Extract hyperslab of selected coordinates}
\usage{
hyperslab_spatial_area(this.coord.file, min.lon, max.lon, min.lat, max.lat)
}
\arguments{
\item{this.coord.file}{}

\item{min.lon}{}

\item{max.lon}{}

\item{min.lat}{}

\item{max.lat}{}
}
\description{
Extract hyperslab of selected coordinates
}
