% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-inegigc.R
\docType{data}
\name{inegigc}
\alias{inegigc}
\title{DATASET_TITLE}
\format{
A data frame with 34081 rows and 11 variables:
\describe{
\item{\code{NOM_ENT}}{character COLUMN_DESCRIPTION}
\item{\code{NOM_MUN}}{character COLUMN_DESCRIPTION}
\item{\code{NOM_LOC}}{character COLUMN_DESCRIPTION}
\item{\code{CVE_ENT}}{character COLUMN_DESCRIPTION}
\item{\code{CVE_MUN}}{character COLUMN_DESCRIPTION}
\item{\code{CVE_LOC}}{character COLUMN_DESCRIPTION}
\item{\code{POBTOT}}{double COLUMN_DESCRIPTION}
\item{\code{POBFEM}}{double COLUMN_DESCRIPTION}
\item{\code{POBMAS}}{double COLUMN_DESCRIPTION}
\item{\code{deci_lon}}{double COLUMN_DESCRIPTION}
\item{\code{deci_lat}}{double COLUMN_DESCRIPTION}
}
}
\usage{
inegigc
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
