% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_clim.R
\name{interp_clim}
\alias{interp_clim}
\title{Add ESM interpolated delta to GLORYS climatology}
\usage{
interp_clim(
  this.rel,
  esm.inter.files,
  glorys.clim.files,
  delta.dir,
  type,
  timestep
)
}
\arguments{
\item{this.rel}{}

\item{esm.inter.files}{}

\item{glorys.clim.files}{}

\item{delta.dir}{}

\item{type}{}

\item{timestep}{}
}
\description{
Add ESM interpolated delta to GLORYS climatology
}
