% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_files.R
\name{merge_files}
\alias{merge_files}
\title{Merge nc files}
\usage{
merge_files(
  this.rel,
  esm.slab.files,
  esm.slab.future.files,
  merged.dir,
  data.type,
  data.period,
  data.range
)
}
\arguments{
\item{esm.slab.files}{}

\item{merged.dir}{}

\item{this.realization}{}
}
\description{
Merge nc files
}
