% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_glorys.R
\name{merge_glorys}
\alias{merge_glorys}
\title{Merge GLORYS files}
\usage{
merge_glorys(glorys.monthly.files, merged.dir, glorys.name)
}
\arguments{
\item{glorys.monthly.files}{}

\item{merged.dir}{}

\item{glorys.name}{}
}
\description{
Merge GLORYS files
}
