% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_permits.R
\name{extract_permits}
\alias{extract_permits}
\title{Extract permit information}
\usage{
extract_permits(
  thispermitfile.no,
  permit.files,
  inegi.cost.comm.all,
  coast.selecc,
  locs.pesca,
  georef.locs.gc
)
}
\arguments{
\item{thispermitfile.no}{}

\item{permit.files}{}

\item{coast.selecc}{}

\item{locs.pesca}{}

\item{inegi.cost.comm}{}

\item{locs.inegi.1}{}

\item{locs.inegi.2}{}

\item{locs.google}{}
}
\value{
this.permit.inegi
}
\description{
Extract permit with location information to match with catch records
}
