% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georef_catch_data.R
\name{georef_catch_data}
\alias{georef_catch_data}
\title{Georeferrence all catch data}
\usage{
georef_catch_data(georefcatchlocs, catchdata)
}
\arguments{
\item{catchdata}{}

\item{pac.buffer}{}
}
\value{
georefcatchlocs
}
\description{
Georeferrence all catch data
}
