% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/georeff_catch_data.R
\name{georref_catch_data}
\alias{georref_catch_data}
\title{Georeferrence all catch data}
\usage{
georref_catch_data(catchdata, georeffcatchlocs, pac.buffer)
}
\arguments{
\item{catchdata}{}

\item{pac.buffer}{}
}
\value{
georeffcatchlocs
}
\description{
Georeferrence all catch data
}
