% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locations_catch.R
\name{get_locations_catch}
\alias{get_locations_catch}
\title{Get fishery office from catch records}
\usage{
get_locations_catch(
  this.catchfileno,
  catch.files,
  coast.selecc,
  inegigc,
  locs.pesca
)
}
\arguments{
\item{catch.files}{catch files}

\item{coast.selecc}{selected coastal states}

\item{inegigc}{all communities in Gulf of California coastal states}

\item{locs.pesca}{corrected location names}
}
\value{
catch.file.rnp
}
\description{
Get fishery office from catch records
}
