% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nearestpointtemp.R
\name{get_nearestpointtemp}
\alias{get_nearestpointtemp}
\title{Obtain temp values for catch}
\usage{
get_nearestpointtemp(
  this.time.interval,
  this.depth,
  catch.pacific.spatial,
  glorys.files,
  catch.geo,
  start.time.interval,
  out.dir,
  no.temps
)
}
\arguments{
\item{this.time.interval}{}

\item{this.depth}{}

\item{catch.pacific.spatial}{}

\item{glorys.files}{}

\item{catch.geo}{}

\item{start.time.interval}{}
}
\description{
Obtain temp values for catch
}
