% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_catch_records.R, R/get_rnpa_catch.R
\name{get_rnpa_catch}
\alias{get_rnpa_catch}
\title{Extract RNPA codes from catch files}
\usage{
get_rnpa_catch(
  this.catchfileno,
  catch.files,
  coast.selecc,
  locs.pesca,
  permit.locs.geo,
  inegi.cost.comm.all,
  georef.locs.gc
)

get_rnpa_catch(
  this.catchfileno,
  catch.files,
  coast.selecc,
  locs.pesca,
  permit.locs.geo,
  inegi.cost.comm.all,
  georef.locs.gc
)
}
\arguments{
\item{catch.files}{catch files}

\item{coast.selecc}{selected coastal states}

\item{locs.pesca}{corrected location names}

\item{permit.locs.geo}{geolocated permits}

\item{inegi.cost.comm.all}{INEGI coastal communities}
}
\value{
catch.file.rnp

catch.file.rnp
}
\description{
Extract RNPA codes from catch files

Extract RNPA codes from catch files
}
