% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_delta.R
\name{interpolate_delta}
\alias{interpolate_delta}
\title{Interpolate delta anomaly}
\usage{
interpolate_delta(this.delta.file, glorys.nc, delta.path)
}
\arguments{
\item{this.delta.file}{}

\item{glorys.nc}{}
}
\value{
inter.raster
}
\description{
Interpolate delta anomaly
}
